refactor: 마커 컴포넌트 분리 #26
